Supporting the debug target

+++ src/functions.c
@@ -595,7 +595,7 @@ static inline unsigned int longest_name_functions(void) {
 
 void print_functions(void) {
     const char *const header = "Name";
-    const unsigned int max_length = max_uint(longest_name_functions(), strlen(header));
+    const unsigned int max_length = max_uint(longest_name_functions(), (unsigned int)strlen(header));
     printf("List of built-in functions:\n");
     printf("%-*s Description\n", max_length, header);
     for (int i = 0; i < functions_quantity; i++) {
