+++ src/main.c
@@ -92,11 +92,11 @@ static void arguments_usage(const char *const parameter) {
     (void)parameter;
     unsigned int cmd_max_length = 0;
     printf("[Usage] %s [Options]\n", software);
-    for (size_t arg_idx = 0; arg_idx < arg_num; arg_idx++) {
+    for (int arg_idx = 0; arg_idx < arg_num; arg_idx++) {
         cmd_max_length = max_uint(cmd_max_length, (unsigned int)strlen(arg_list[arg_idx].cmd));
     }
     printf("[Options]:\n");
-    for (size_t arg_idx = 0; arg_idx < arg_num; arg_idx++) {
+    for (int arg_idx = 0; arg_idx < arg_num; arg_idx++) {
         printf("\t%-*s or -%c: %s\n", cmd_max_length, arg_list[arg_idx].cmd, arg_list[arg_idx].cmd[2], arg_list[arg_idx].usage);
     }
     actions |= ACTION_EXIT;
